code/python/test-platform/test_platform/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2020.4.5.1
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
colorlog==5.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
gevent==21.1.2
greenlet==1.1.0
har2case==0.3.1
HttpRunner==1.5.8
idna==2.9
iteration==1.2.0
itypes==1.2.0
Jinja2==3.0.1
lxml==4.5.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
openpyxl==3.0.7
paramiko==2.7.2
pathlib==1.0.1
Pillow==8.0.1
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.9
view==0.1
zope.event==4.5.0
zope.interface==5.4.0
